pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "destill3d"
version = "0.1.0"
description = "3D model feature extraction and classification toolkit"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "LayerDynamics / Lattice Labs" }
]
keywords = ["3d", "point-cloud", "classification", "feature-extraction", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Graphics :: 3D Modeling",
]
dependencies = [
    "numpy>=1.24",
    "trimesh>=4.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sqlalchemy>=2.0",
    "typer>=0.9",
    "rich>=13.0",
    "protobuf>=4.0",
    "httpx>=0.25",
    "pyyaml>=6.0",
    "scipy>=1.11",
    "scikit-learn>=1.0",
    "tenacity>=8.0",
    "aiofiles>=23.0",
    "keyring>=24.0",
]

[project.optional-dependencies]
cad = [
    "pythonocc-core>=7.7",
]
open3d = [
    "open3d>=0.17",
]
gpu = [
    "torch>=2.0",
    "onnxruntime-gpu>=1.16",
]
cpu = [
    "torch>=2.0",
    "onnxruntime>=1.16",
]
faiss = [
    "faiss-cpu>=1.7",
]
faiss-gpu = [
    "faiss-gpu>=1.7",
]
acquire = [
    "aiofiles>=23.0",
    "keyring>=24.0",
]
zero-shot = [
    "torch>=2.0",
    "transformers>=4.30",
]
rendering = [
    "pyrender>=0.1.45",
    "pyglet>=2.0",
]
server = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "python-multipart>=0.0.6",
]
hdf5 = [
    "h5py>=3.9",
]
tfrecord = [
    "tensorflow>=2.14",
]
scraping = [
    "beautifulsoup4>=4.12",
    "lxml>=4.9",
]
lidar = [
    "laspy>=2.4",
]
postgresql = [
    "psycopg2-binary>=2.9",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1",
]
full = [
    "destill3d[cad,gpu,faiss-gpu,postgresql,acquire,zero-shot,rendering,hdf5,server]",
]

[project.scripts]
destill3d = "destill3d.cli.main:app"

[project.urls]
Homepage = "https://github.com/layerdynamics/destill3d"
Documentation = "https://github.com/layerdynamics/destill3d#readme"
Repository = "https://github.com/layerdynamics/destill3d"
Issues = "https://github.com/layerdynamics/destill3d/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/destill3d"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
